fixed errors

+++ jedi_raycaster_build/Map.cpp
@@ -48,7 +48,7 @@ bool Map::isOutSideMap(float x, float y)
 void Map::renderMapGrid(olc::PixelGameEngine* PGEptr)
 {
     // fill background for minimap
-    PGEptr->FillRect( 0, 0, (int)MAP_NUM_COLS_X * TILE_SIZE * MINIMAP_SCALE_FACTOR, (int)MAP_NUM_ROWS_Y * TILE_SIZE * MINIMAP_SCALE_FACTOR, olc::DARK_YELLOW );
+    PGEptr->FillRect( 0, 0, (int)(MAP_NUM_COLS_X * TILE_SIZE * MINIMAP_SCALE_FACTOR), (int)(MAP_NUM_ROWS_Y * TILE_SIZE * MINIMAP_SCALE_FACTOR), olc::DARK_YELLOW );
 
     // draw each tile
 	for (int i = 0; i < MAP_NUM_ROWS_Y; i++) {
@@ -67,10 +67,10 @@ void Map::renderMapGrid(olc::PixelGameEngine* PGEptr)
 			}
             // render this thile
 			PGEptr->FillRect(
-				tileX * MINIMAP_SCALE_FACTOR     + 1,
-				tileY * MINIMAP_SCALE_FACTOR     + 1,
-				TILE_SIZE * MINIMAP_SCALE_FACTOR - 1,
-				TILE_SIZE * MINIMAP_SCALE_FACTOR - 1,
+				(int)(tileX * MINIMAP_SCALE_FACTOR     + 1),
+				(int)(tileY * MINIMAP_SCALE_FACTOR     + 1),
+				(int)(TILE_SIZE * MINIMAP_SCALE_FACTOR - 1),
+				(int)(TILE_SIZE * MINIMAP_SCALE_FACTOR - 1),
 				p
 			);
 		}
